backend/pyproject.toml
[project]
name = "accountabilidash"
version = "0.1.0"
description = "Accountabilidash - Backend API"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.115.0",
    "sqlmodel>=0.0.22",
    "alembic>=1.14.0",
    "asyncpg>=0.30.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "pydantic-settings>=2.7.0",
    "bcrypt>=4.0.0",
    "python-jose[cryptography]>=3.3.0",
    "structlog>=24.4.0",
    "psycopg2-binary>=2.9.0",
]

[dependency-groups]
dev = [
    "ruff>=0.9.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.0",
    "aiosqlite>=0.20.0",
]

[tool.ruff]
target-version = "py313"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM", "RUF"]
# B008: Depends() in defaults is the canonical FastAPI DI pattern.
per-file-ignores = { "app/routers/*" = ["B008"], "app/core/deps.py" = ["B008"] }

[tool.pytest.ini_options]
asyncio_mode = "auto"
